# server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VAIPE API",
    "description": "VAIPE API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/oxy": {
      "get": {
        "tags": [
          "API Gateway - Oxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appid",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Gateway - Oxy"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appid",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "number",
                  "example": 99.5
                },
                "signature": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "example": "any"
                  },
                  "signature": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/medicineReceipt": {
      "get": {
        "tags": [
          "API Gateway - Medicine Receipt"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appid",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Gateway - Medicine Receipt"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appid",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "ecg demo"
                },
                "drugs": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Flu drug"
                      },
                      "quantity": {
                        "type": "number",
                        "example": 5
                      },
                      "price": {
                        "type": "number",
                        "example": 10000
                      }
                    }
                  }
                },
                "signature": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "drugs": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "signature": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "API Gateway - Medicine Receipt"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appid",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiptIds": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}